# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kalshi BTC Bot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MarketInfo {
  ticker: string;
  expirationTime: string;
  timeLeftSeconds: number;
  yesPrice: number | null;
  noPrice: number | null;
}

export type BotStatusStatus =
  (typeof BotStatusStatus)[keyof typeof BotStatusStatus];

export const BotStatusStatus = {
  idle: "idle",
  running: "running",
  stopped: "stopped",
  error: "error",
} as const;

export interface Settlement {
  ticker: string;
  side: "yes" | "no";
  result: "yes" | "no";
  won: boolean;
  buyPriceCents: number;
  totalBought: number;
  soldViaStopLoss: number;
  settledCount: number;
  stopLossPnlCents: number;
  settlementPnlCents: number;
  totalPnlCents: number;
  settledAt: string;
}

export interface StopLossTier {
  /** Trigger price in cents (e.g. 90 = $0.90) */
  priceCents: number;
  /** Fraction of total position to sell at this tier (e.g. 0.333 = 1/3) */
  fraction: number;
}

export interface BotStatus {
  status: BotStatusStatus;
  startedAt: string | null;
  currentMarket: MarketInfo | null;
  lastPollAt: string | null;
  lastError: string | null;
  totalTrades: number;
  tradedMarkets: string[];
  totalPnlCents: number;
  winCount: number;
  lossCount: number;
  accuracy: number | null;
  useStopLoss: boolean;
  stopLossTiers: StopLossTier[];
}

export type TradeSide = (typeof TradeSide)[keyof typeof TradeSide];

export const TradeSide = {
  yes: "yes",
  no: "no",
} as const;

export interface Trade {
  id: string;
  ticker: string;
  side: TradeSide;
  price: number;
  count: number;
  timestamp: string;
}

export interface TradeList {
  trades: Trade[];
}

export interface BotStartConfig {
  /** Number of contracts to buy per trade (ignored when useAllBalance is true) */
  tradeSize?: number;
  /** When true, stake the entire account balance on each trade instead of a fixed contract count */
  useAllBalance?: boolean;
  /** Price threshold in cents (e.g. 97 = $0.97) */
  thresholdCents?: number;
  /** Seconds before expiry to start watching the market */
  windowSeconds?: number;
  /** Polling interval in milliseconds */
  checkIntervalMs?: number;
  /** Whether to enable stop loss selling */
  useStopLoss?: boolean;
  /** Ordered stop loss tiers (last tier always sells all remaining) */
  stopLossTiers?: StopLossTier[];
  /** EMA smoothing factor 0-1 (default 0.2) */
  emaAlpha?: number;
  /** EMA must reach this value before entry fires (default 0.88) */
  emaThreshold?: number;
}

export interface BotActionResponse {
  success: boolean;
  message: string;
}