# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Kalshi BTC Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bot/start:
    post:
      operationId: startBot
      tags: [bot]
      summary: Start the bot
      description: Start the bot with optional configuration overrides
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotStartConfig"
      responses:
        "200":
          description: Bot started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotActionResponse"
        "400":
          description: Error starting bot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotActionResponse"
...
components:
  schemas:
...
    BotStartConfig:
      type: object
      properties:
        tradeSize:
          type: integer
          description: Number of contracts to buy per trade (ignored when useAllBalance is true)
        useAllBalance:
          type: boolean
          description: When true, stake the entire account balance on each trade instead of a fixed contract count
        thresholdCents:
          type: integer
          description: Price threshold in cents (e.g. 99 = $0.99)
        windowSeconds:
          type: integer
          description: Seconds before expiry to start watching
        checkIntervalMs:
          type: integer
          description: Polling interval in milliseconds
        useStopLoss:
          type: boolean
          description: Whether to enable stop-loss tiers
        stopLossTiers:
          type: array
          items:
            $ref: "#/components/schemas/StopLossTier"
          description: Stop-loss tier definitions
        emaAlpha:
          type: number
          description: EMA smoothing factor (0-1, default 0.2)
        emaThreshold:
          type: number
          description: EMA must reach this value before entry fires (default 0.88)
    BotActionResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message